QSaveFile: code tidies
Turn a cast into a call to QFlags::toInt(), plus a cast.

This is in preparation for a future patch that will remove the
implicit conversion. So why not casting directly? Because we don't
know if the QFlags is backed by int or uint, so using the wrong
cast won't compile; and we MUST pass unsigned because of the %x
specifier. Granted, one could've used a static_assert, but I prefer
a more flexible and idiomatic solution.

+++ src/corelib/io/qsavefile.cpp
@@ -204,7 +204,7 @@ bool QSaveFile::open(OpenMode mode)
     // In the future we could implement ReadWrite by copying from the existing file to the temp file...
     // The implications of NewOnly and ExistingOnly when used with QSaveFile need to be considered carefully...
     if (mode & (ReadOnly | Append | NewOnly | ExistingOnly)) {
-        qWarning("QSaveFile::open: Unsupported open mode 0x%x", int(mode));
+        qWarning("QSaveFile::open: Unsupported open mode 0x%x", uint(mode.toInt()));
         return false;
     }
